# Backend/myapp/static/swagger.yaml
openapi: 3.0.0

info:
  title: "AutoMate App API"
  description: "An API for obtaining existing information about successful applications of robotics solutions in the industry and for posting data about new use cases."
  version: "1.0.0"
  contact:
    name: Aleksandra Rubaj
    email: ola@yway.pl

#servers:

paths:
  /use-cases:
    get:
      tags: [Use cases]
      summary: Returns a list of use-cases.
      operationId: getUseCases
      # query parameters
      parameters:
        - name: bodyLimit
          in: query
          description: The amount of use-cases returned.
          schema:
            type: integer
            minimum: 1
            maximum: 15
            default: 10
            example: 10
        - name: industries
          in: query
          description: Robotics implementations from specified industry branches.
          schema:
            type: array
            items:
              type: string
          explode: false
          example:
            [
              "Food and Agriculture",
              "Electronics and Technology",
              "Furniture and Equipment",
            ]
        - name: applications
          in: query
          description: Robotics implementations by specified applications.
          schema:
            type: array
            items:
              type: string
          explode: false
          example: ["Pick and Place", "Machine Tending", "Welding"]
        - name: country
          in: query
          description: Robotics implementations by country.
          schema:
            type: array
            items:
              type: string
          example: ["Czech Republic", "Poland", "Spain"]

      responses:
        "200":
          description: A JSON array of use-cases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseCases"
        default:
          description: Unexpected error.
    post:
      tags: [Use cases]
      summary: Creates a new use-case in the database.
      operationId: addUseCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseCase"

      responses:
        "200":
          description: Succesfully created a new use-case.

  /use-cases/{caseId}:
    summary: Represents a single use-case.
    description: >
      This resource represents a specific use-case in the system.
      Each use-case is identified by a numeric 'id'.
    get:
      tags: [Use cases]
      summary: Returns a use case by ID.
      operationId: getUseCase
      #query parameters
      parameters:
        - name: caseId
          in: path
          description: Numeric ID of the use case to get.
          required: true
          schema:
            type: integer
            example: 7

      responses:
        "200":
          description: Successful pull of the use-case info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseCase"
        "400":
          description: The specified use-case ID is invalid (not a number).
        "404":
          description: A use-case with the specified ID was not found.
        default:
          description: Unexpected error.

components:
  schemas:
    UseCases:
      description: Array of use-cases
      type: array
      items:
        $ref: "#/components/schemas/UseCase"

    UseCase:
      type: object
      description: Model of use-case info
      properties:
        id:
          type: string
          example: "5e63f7ecbcfe1b0c14f28c69"
          readOnly: true
        filter_tags:
          type: object
          properties:
            Applications:
              type: array
              items:
                type: string
              example: ["Machine Tending", "Pick and Place"]
            Company:
              type: string
              example: "2K Trend, a.s."
            Industry:
              type: string
              example: "Plastic molding"
            Country:
              type: string
              example: "Czech Republic"
            Company size:
              type: string
              example: "30+"
          required: [Applications, Company, Industry]
        content:
          type: object
          properties:
            Article_title:
              type: string
            Article_sections:
              type: object
            Bullet_points:
              type: object
          required: [Article_title, Article_sections]
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: url
              path:
                type: string
              checksum:
                type: string

#security

tags:
  - name: Use cases
    description: Applications of robotics solutions
  - name: Knowledge
    description: Robotics tutorials, courses and articles
#externalDocs